+++ eval/lib/oauth2.go
@@ -46,8 +46,7 @@ func NewOAuthAuthorizationUrlFunction(ctx *hcl.EvalContext, oauth2Configs []conf
 				return emptyStringVal, fmt.Errorf("undefined reference: %s", label)
 			}
 
-			uid := seetie.ToString(seetie.ValueToMap(ctx.Variables["request"])["id"])
-			authorizationEndpoint, err := oauth2.GetAuthorizationEndpoint(uid)
+			authorizationEndpoint, err := oauth2.GetAuthorizationEndpoint()
 			if err != nil {
 				return emptyStringVal, err
 			}
@@ -89,7 +88,7 @@ func NewOAuthAuthorizationUrlFunction(ctx *hcl.EvalContext, oauth2Configs []conf
 				query.Set("scope", scope)
 			}
 
-			verifierMethod, err := oauth2.GetVerifierMethod(uid)
+			verifierMethod, err := oauth2.GetVerifierMethod()
 			if err != nil {
 				return cty.StringVal(""), err
 			}