+++ handler/transport/backend.go
@@ -216,7 +216,7 @@ func (b *Backend) RoundTrip(req *http.Request) (*http.Response, error) {
 func (b *Backend) openAPIValidate(req *http.Request, tc *Config, deadlineErr <-chan error) (*http.Response, error) {
 	requestValidationInput, err := b.openAPIValidator.ValidateRequest(req)
 	if err != nil {
-		return nil, errors.BackendValidation.Label(b.name).Kind("backend_request_validation").With(err)
+		return nil, errors.BackendOpenapiValidation.Label(b.name).With(err)
 	}
 
 	beresp, err := b.innerRoundTrip(req, tc, deadlineErr)
@@ -225,8 +225,7 @@ func (b *Backend) openAPIValidate(req *http.Request, tc *Config, deadlineErr <-c
 	}
 
 	if err = b.openAPIValidator.ValidateResponse(beresp, requestValidationInput); err != nil {
-		return nil, errors.BackendValidation.Label(b.name).Kind("backend_response_validation").
-			With(err).Status(http.StatusBadGateway)
+		return nil, errors.BackendOpenapiValidation.Label(b.name).With(err).Status(http.StatusBadGateway)
 	}
 
 	return beresp, nil
@@ -283,19 +282,17 @@ func (b *Backend) innerRoundTrip(req *http.Request, tc *Config, deadlineErr <-ch
 	return beresp, nil
 }
 
-const backendTokenRequest = "backendTokenRequest"
-
 func (b *Backend) withTokenRequest(req *http.Request) error {
 	if b.tokenRequest == nil {
 		return nil
 	}
 
-	trValue, _ := req.Context().Value(backendTokenRequest).(string)
+	trValue, _ := req.Context().Value(request.BackendTokenRequest).(string)
 	if trValue != "" { // prevent loop
 		return nil
 	}
 
-	ctx := context.WithValue(req.Context(), backendTokenRequest, "tr")
+	ctx := context.WithValue(req.Context(), request.BackendTokenRequest, "tr")
 	// Reset for upstream transport; prevent mixing values.
 	// tokenRequest will have their own backend configuration.
 	ctx = context.WithValue(ctx, request.BackendParams, nil)
@@ -309,7 +306,7 @@ func (b *Backend) withRetryTokenRequest(req *http.Request, res *http.Response) (
 		return false, nil
 	}
 
-	trValue, _ := req.Context().Value(backendTokenRequest).(string)
+	trValue, _ := req.Context().Value(request.BackendTokenRequest).(string)
 	if trValue != "" { // prevent loop
 		return false, nil
 	}